IFA/requirements.txt
ace_tools==0.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.45.2
anyio==4.8.0
asttokens==3.0.0
async-timeout==4.0.3
attrs==25.1.0
Authlib==1.3.1
beautifulsoup4==4.13.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cryptography==44.0.1
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.11
decorator==5.1.1
distro==1.9.0
duckduckgo_search==7.3.2
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.8
filelock==3.17.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.2.0
greenlet==3.1.1
grpcio==1.70.0
grpcio-health-checking==1.70.0
grpcio-tools==1.70.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.28.1
ibm-cos-sdk==2.13.6
ibm-cos-sdk-core==2.13.6
ibm-cos-sdk-s3transfer==2.13.6
ibm_watsonx_ai==1.2.8
idna==3.10
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==8.32.0
isodate==0.7.2
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
langchain==0.3.19
langchain-anthropic==0.3.7
langchain-community==0.3.18
langchain-core==0.3.37
langchain-ibm==0.3.6
langchain-text-splitters==0.3.6
langsmith==0.3.8
litellm==1.61.6
lomond==0.3.3
lxml==5.3.1
markdown-it-py==3.0.0
markdownify==0.14.1
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.1.0
mypy-extensions==1.0.0
nest_asyncio==1.6.0
numexpr==2.10.2
numpy==1.26.4
openai==1.63.0
orjson==3.10.15
packaging==24.2
pandas==2.1.4
parso==0.8.4
pickleshare==0.7.5
pillow==11.1.0
pip==25.0
platformdirs==4.3.6
primp==0.12.1
prompt_toolkit==3.0.50
propcache==0.2.1
protobuf==5.29.3
psutil==5.9.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydantic-settings==2.8.0
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
pywin32==308
PyYAML==6.0.2
pyzmq==24.0.1
rdflib==7.1.3
referencing==0.36.2
regex==2024.11.6
requests==2.32.2
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
scikit-learn==1.6.1
scipy==1.15.2
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SPARQLWrapper==2.0.0
SQLAlchemy==2.0.38
stack_data==0.6.3
starlette==0.45.3
tabulate==0.9.0
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.9.0
tokenizers==0.21.0
tornado==6.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
typing-inspect==0.9.0
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
validators==0.34.0
wcwidth==0.2.13
weaviate-client==4.10.4
wheel==0.45.1
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
